requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
aniso8601==9.0.1
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
cloudinary==1.38.0
decorator==5.1.1; python_version >= '3.7'
executing==2.0.1; python_version >= '3.5'
faker==22.2.0; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-cors==4.0.0
flask-migrate==4.0.5; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.0.3; python_version >= '3.7'
gunicorn==21.2.0; python_version >= '3.5'
importlib-metadata==7.0.1; python_version < '3.10'
importlib-resources==6.1.1; python_version < '3.9'
ipdb==0.13.9; python_version >= '2.7'
ipython==8.12.3; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.3; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.25; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
stack-data==0.6.3
toml==0.10.2; python_version >= '3.7'
traitlets==5.14.1; python_version >= '3.8'
typing-extensions==4.9.0; python_version < '3.9'
urllib3==2.1.0; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
